Add "Mesh to Strokes" tool to allow working with original strokes  (#10)
* install Burst

Burst is being installed because it will be used to make the "Mesh to strokes" tool fast

* Add MathematicsUtils.cs to help with Unity.Mathematics

* implement initial functionality for "Mesh to Strokes" tool

Todo:

- optimize, e.g if you have multiple meshes selected and you use the tool, it'll run MeshToStroke() for each of them synchronously, one by one.

- fix bugs. e.g some brushes scale gets messed up when using this tool.

* Add progress bar

* fix name of Undo operation

* optimize MeshToStrokes's asset creation

The AssetDatabase.StartAssetEditing() and AssetDatabase.StopAssetEditing() methods disable importing of assets that we create in GetMeshSubset, and we instead do it all at once after the assets have been created. The performance gain is <= 2x

* make sure submesh assets are saved to a common folder

* add proper error handling to MeshToStrokes

* make sure new meshes also get the timestamp data

the new meshes need the uv2 timestamp data if you wanna use it in e.g shaders.

* fix: make sure only a subset of uv2 gets copied to new mesh

* fix uv stuff

This commit does 2 things:

1. change variable naming to the correct timestamp uv channel: uv2 -> uv3 (texcoord2)

2. make sure both uv2 and uv3 are copied to the new meshes

(2) fixes issue where e.g Dots shader doesn't work because uv2 data doesn't get copied to the new mesh

* Implement method for safe copying of mesh attributes

* make sure that mesh assets have unique names

* add proper error handling for no timestamp data

* fix issue with modified transforms

This commit fixes an issue where if the user has modified the transform of a mesh they are separating (e.g scaled it by 2), the 'Mesh To Strokes' tool wouldn't create the new meshes with the 2x scale taken into account. The resulting meshes would be the original size

* change name MeshToStrokes -> SplitMeshByStroke

We decided to change the name as this is more accurate for what the tool does.

* Change name SeparateStrokesByBrushColor -> SplitMeshByColor

This name describes the tool more accurately